[FEAT] intersection observe 기능 분리
